requirements.txt
absl-py                  == 0.15.0
argcomplete              == 1.12.3
argon2-cffi              == 21.1.0
asn1crypto               == 0.24.0
astunparse               == 1.6.3
async-generator          == 1.10
attrs                    == 21.2.0
autopep8                 == 1.6.0
backcall                 == 0.2.0
bleach                   == 4.1.0
Brotli                   == 1.0.9
cached-property          == 1.5.2
cachetools               == 4.2.2
certifi                  == 2021.10.8
cffi                     == 1.15.0
chardet                  == 4.0.0
charset-normalizer       == 2.0.7
click                    == 8.0.3
colorama                 == 0.4.4
cycler                   == 0.11.0
cytoolz                  == 0.11.0
dash                     == 2.0.0
dash-bootstrap-components== 1.0.0
dash-core-components     == 2.0.0
dash-html-components     == 2.0.0
dash-table               == 5.0.0
dataclasses              == 0.6
debugpy                  == 1.5.1
decorator                == 4.4.2
defusedxml               == 0.7.1
entrypoints              == 0.3
Flask                    == 2.0.2
Flask-Compress           == 1.10.1
flatbuffers              == 1.12
fonttools                == 4.27.1
gast                     == 0.4.0
google-auth              == 2.3.3
google-auth-oauthlib     == 0.4.6
google-pasta             == 0.2.0
grpcio                   == 1.34.1
h5py                     == 3.1.0
idna                     == 3.3
imageio                  == 2.11.0
importlib-metadata       == 4.8.1
importlib-resources      == 5.4.0
ipykernel                == 5.5.6
ipython                  == 7.16.1
ipython-genutils         == 0.2.0
itsdangerous             == 2.0.1
jedi                     == 0.18.0
jellyfish                == 0.8.2
Jinja2                   == 3.0.2
joblib                   == 1.0.1
jsonschema               == 3.2.0
jupyter-client           == 7.0.6
jupyter-core             == 4.9.1
jupyterlab-pygments      == 0.1.2
keras-nightly            == 2.5.0.dev2021032900
Keras-Preprocessing      == 1.1.2
keyring                  == 10.6.0
keyrings.alt             == 3.0
kiwisolver               == 1.3.1
Markdown                 == 3.3.4
MarkupSafe               == 2.0.1
matplotlib               == 3.3.4
matplotlib-inline        == 0.1.3
mistune                  == 0.8.4
nbclient                 == 0.5.5
nbconvert                == 6.0.7
nbformat                 == 5.1.3
nest-asyncio             == 1.5.1
networkx                 == 2.5.1
notebook                 == 6.4.5
numpy                    == 1.19.5
oauthlib                 == 3.1.1
opt-einsum               == 3.3.0
packaging                == 21.2
pandas                   == 1.1.5
pandocfilters            == 1.5.0
parso                    == 0.8.2
pathy                    == 0.5.2
pexpect                  == 4.8.0
pickleshare              == 0.7.5
Pillow                   == 8.4.0
pip                      == 21.3.1
plotly                   == 5.3.1
prometheus-client        == 0.12.0
prompt-toolkit           == 3.0.22
protobuf                 == 3.19.1
ptyprocess               == 0.7.0
pyasn1                   == 0.4.8
pyasn1-modules           == 0.2.8
pycodestyle              == 2.8.0
pycparser                == 2.21
Pygments                 == 2.10.0            
pyparsing                == 2.4.7
Pyphen                   == 0.10.0
pyrsistent               == 0.18.0
python-dateutil          == 2.8.2
pytz                     == 2021.3
PyWavelets               == 1.1.1
pyxdg                    == 0.25
pyzmq                    == 22.3.0
requests                 == 2.26.0
requests-oauthlib        == 1.3.0
rsa                      == 4.7.2
scikit-image             == 0.17.2
scipy                    == 1.5.4
SecretStorage            == 2.3.1
Send2Trash               == 1.8.0
setuptools               == 56.2.0
setuptools-scm           == 6.3.2
six                      == 1.15.0
smart-open               == 3.0.0
spacy-legacy             == 3.0.5
tenacity                 == 8.0.1
tensorboard              == 2.7.0
tensorboard-data-server  == 0.6.1
tensorboard-plugin-wit   == 1.8.0
tensorflow               == 2.5.0
tensorflow-estimator     == 2.5.0
termcolor                == 1.1.0
terminado                == 0.12.1
testpath                 == 0.5.0
threadpoolctl            == 2.1.0
tifffile                 == 2020.9.3
toml                     == 0.10.2
tomli                    == 1.2.2
toolz                    == 0.11.1
tornado                  == 6.1
tqdm                     == 4.62.3
traitlets                == 4.3.3
typer                    == 0.4.0
typing-extensions        == 3.7.4.3
urllib3                  == 1.26.7
wcwidth                  == 0.2.5
webencodings             == 0.5.1
Werkzeug                 == 2.0.2
wheel                    == 0.36.2
wincertstore             == 0.2
wrapt                    == 1.12.1
zipp                     == 3.6.0